browser class now using mixins
