+++ lib/browser_core.js
@@ -1,0 +1,9 @@
+"use strict";
+
+
+module.exports = class BrowserCore {
+    constructor(page, settings) {
+        this.page = page;
+        this._settings = settings;
+    }
+};
