+++ lib/wendigo.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const process = require('process');
-const Browser = require('./browser');
+const BrowserFactory = require('./browser_factory');
 const puppeteer = require('puppeteer');
 const Errors = require('./errors/errors');
 
@@ -19,7 +19,7 @@ module.exports = class Wendigo {
         settings = Object.assign(defaultSettings, settings);
         await this._setInstance(settings);
         const page = await this.instance.newPage();
-        return new Browser(page, settings);
+        return BrowserFactory.createBrowser(page, settings);
     }
 
     static async stop() {
